Add sandboxing with Landlock
p7zip handles a lot of archive formats, each bringing complexity and extending
its attack surface.  Landlock is a Linux security sandboxing mechanism which
helps limit the impact of bug exploitation.  These modifications enable to use
Landlock following a best-effort appoach: if Landlock is available on the
running system, then it will be used to sandbox p7zip, otherwise nothing will
change.

By default, all console commands (7z, 7za and 7zr) reading archives will create
a sandbox which denies basic file-system accesses (e.g. read, create or write
to files).  All other commands (b, h, i, t) are untouched.

FIXME: For now, only basic use of list (l) and extracts (e, x) commands is
working and tested.

The sandbox is enforced when calling COpenCallbackConsole::Open_setTotal(),
which seems to be the more generic place for this purpose.

When extracting an archive, a new helper called GetOutputBaseDir() is used to
add security exception to the sandbox so that it will be allowed to create
files in the output directory.

Signed-off-by: Mickaël Salaün <mic@digikod.net>